opti shading

+++ includes/cub3d_bonus.h
@@ -6,7 +6,7 @@
 /*   By: teliet <teliet@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/16 12:45:39 by axlamber          #+#    #+#             */
-/*   Updated: 2023/03/21 11:18:15 by teliet           ###   ########.fr       */
+/*   Updated: 2023/03/21 11:29:32 by teliet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,7 +83,7 @@ int					get_b(int trgb);
 int					create_trgb(int t, int r, int g, int b);
 
 // Color operations
-int				add_shade(int trgb, float shade);
+int				add_shade(int trgb, int shade);
 
 // Ray casting
 t_collision		cast_2D_ray(t_game *game, t_vector3d direction);
